осталось потестировать перед показом

+++ movemaker.h
@@ -15,7 +15,7 @@ public:
     Q_INVOKABLE void move();
     bool hasMove() const;
     void moveCompleted();
-    bool updateCache(const qint64 newValue);
+    bool updateCache(const quint16 newValue);
 
 signals:
     void noMoreMoves();
@@ -23,7 +23,7 @@ signals:
 private:
     int m_patternSize;
     bool m_dirtyMove;
-    QSet<qint64> m_imageCaches;
+    QSet<quint16> m_imageCaches;
 };
 
 #endif // MOVEMAKER_H
